Criação do banco e conexões

+++ GerenciadorEstoque/src/model/Produto.java
@@ -4,11 +4,11 @@ public class Produto {
     //definir variaveis
     private int id, quantity;
     private String name;
-    private float price;
+    private double price;
 
 
     //contrutor
-    public Produto (int id, String name, float price, int quantity) {
+    public Produto (int id, String name, double price, int quantity) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -30,10 +30,10 @@ public class Produto {
         this.name = name;
     }
 
-    public float getPrice(){
+    public double getPrice(){
         return price;
     }
-    public void setPrice(float price){
+    public void setPrice(double price){
         this.price = price;
     }
